Removed advent of code.
The advent of code problems will be tackled as part of this repo when time permits, but through a separate test project. Given that only the first part of the first problem for 2023 was solved, I'm okay scrapping the entire setup for it until I have time to move on it again.